+++ src/modules/project/forms/ProjectCreationWalletConnectionForm.tsx
@@ -4,14 +4,12 @@ import {
   ConnectionOption,
   LightingWalletForm,
   Limits,
-  NWCWalletForm,
 } from '../pages/projectCreation/hooks/useWalletForm'
 
 type Props = {
   readOnly?: boolean
   connectionOption: ConnectionOption
   lightningAddress: LightingWalletForm
-  nwc: NWCWalletForm
   setConnectionOption: (connectionOption: ConnectionOption) => void
   limits: Limits
   removeSponsors?: boolean
@@ -21,7 +19,6 @@ export const ProjectCreationWalletConnectionForm = ({
   readOnly,
   connectionOption,
   lightningAddress,
-  nwc,
   setConnectionOption,
   limits,
   removeSponsors,
@@ -31,7 +28,6 @@ export const ProjectCreationWalletConnectionForm = ({
       readOnly={readOnly}
       connectionOption={connectionOption}
       lightningAddress={lightningAddress}
-      nwc={nwc}
       setConnectionOption={setConnectionOption}
       limits={limits}
       removeSponsors={removeSponsors}