+++ src/modules/project/graphql/mutation/publishNostrEventMutation.ts
@@ -1,7 +1,0 @@
-import { gql } from '@apollo/client'
-
-export const PUBLISH_NOSTR_EVENT_MUTATION = gql`
-  mutation PublishNostrEvent($event: String!) {
-    publishNostrEvent(event: $event)
-  }
-`