+++ src/modules/project/pages/projectFunding/views/fundingSuccess/components/SuccessImageComponent.tsx
@@ -11,7 +11,6 @@ import {
   ContributionSuccessIllustrationUrl,
 } from '@/shared/constants/index.ts'
 
-import { BecomeAnAmbassador } from './BecomeAmbassador.tsx'
 import { ContributionStatusSection } from './ContributionStatusSection.tsx'
 import { SubscribeToGeyser } from './SubscribeToGeyser.tsx'
 
@@ -20,7 +19,6 @@ export const SuccessImageComponent = ({ isPending }: { isPending: boolean }) => 
     <VStack w="full" alignItems="start" spacing={12}>
       <SuccessfulContributionBanner isPending={isPending} />
       <SubscribeToGeyser />
-      <BecomeAnAmbassador />
     </VStack>
   )
 }